+++ docs_snippets/lib/snippets/guides/adding_theme_properties/getter.dart
@@ -1,10 +1,15 @@
+import 'package:flutter/widgets.dart';
+
 import 'package:forui/forui.dart';
 
-import 'package:docs_snippets/snippets/guides/adding_theme_properties/brand_style.dart';
+import 'package:docs_snippets/snippets/guides/adding_theme_properties/brand_colors.dart';
 
 // {@snippet}
-extension BrandStyleExtension on FThemeData {
-  BrandStyle get brand => extension<BrandStyle>();
+extension BrandColorsExtension on FColors {
+  BrandColors get brand => extension<BrandColors>();
+
+  // Alternatively, you can create a getter to access extension fields directly.
+  Color get accent => extension<BrandColors>().accent;
 }
 
 // {@endsnippet}