+++ forui/lib/src/theme/colors.dart
@@ -818,7 +818,14 @@ final class FColors with Diagnosticable {
   /// ```dart
   /// extension FColorsBrandColors on FColors {
   ///   BrandColors get brand => extension<BrandColors>();
+  ///
+  ///   // Alternatively
+  ///   Color get accent => extension<BrandColors>().accent;
   /// }
+  ///
+  /// final brand = context.theme.colors.brand;
+  ///
+  /// final accent = context.theme.colors.accent;
   /// ```
   /// {@endtemplate}
   T extension<T extends Object>() => _extensions[T]! as T;